# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChessCouncil",
    "description": "多 Agent 协作辩论的实时国际象棋分析与对战系统",
    "version": "0.4.1"
  },
  "paths": {
...
    "/api/rooms": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Create Room",
        "operationId": "create_room_api_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}/join": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Join Room",
        "operationId": "join_room_api_rooms__room_id__join_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRoomRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Get Room",
        "operationId": "get_room_api_rooms__room_id__get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}/move": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Room Move",
        "operationId": "room_move_api_rooms__room_id__move_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomMoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}/reset": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Room Reset",
        "operationId": "room_reset_api_rooms__room_id__reset_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateRoomRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 24,
            "title": "Name",
            "default": "玩家"
          },
          "color": {
            "type": "string",
            "enum": [
              "white",
              "black"
            ],
            "title": "Color",
            "default": "white"
          }
        },
        "type": "object",
        "title": "CreateRoomRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JoinRoomRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 24,
            "title": "Name",
            "default": "玩家"
          }
        },
        "type": "object",
        "title": "JoinRoomRequest"
      },
...
      "RoomMoveRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "uci": {
            "type": "string",
            "title": "Uci"
          }
        },
        "type": "object",
        "required": [
          "token",
          "uci"
        ],
        "title": "RoomMoveRequest"
      },
      "RoomResetRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "RoomResetRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}